+++ modules/settings/settings-helpers.js
@@ -315,8 +315,10 @@ export default class SettingsHelpers {
       onChange: this.debouncedReload,
     });
 
-    // auto-configure the default values of tokens - once
-    game.settings.register("starwarsffg", "token_configured", {
+    // Auto-configure some tokens settings on creation
+    game.settings.register("starwarsffg", "RivalTokenPrepend", {
+      name: game.i18n.localize("SWFFG.Settings.actor.RivalTokenPrepend.Name"),
+      hint: game.i18n.localize("SWFFG.Settings.actor.RivalTokenPrepend.Hint"),
       scope: "world",
       config: false,
       default: false,